refactor: changed the location of some components
